refactor: рефактор валидации для форм
