+++ packages/client/src/pages/RegistrationPage/index.tsx
@@ -3,32 +3,9 @@ import { Button, Form, Input, Flex, Typography } from 'antd'
 import { NavLink, useNavigate } from 'react-router-dom'
 import { RouterPaths } from '@/shared/router'
 import { useAuth } from '@/shared/hooks'
+import { fieldTooltip, regExpByField, validate } from '@/shared/lib/Validation'
 
 const { Title } = Typography
-
-export const fieldTooltip = {
-  first_name:
-    'латиница или кириллица, первая буква должна быть заглавной, без пробелов и без цифр, нет спецсимволов (допустим только дефис)',
-  second_name:
-    'латиница или кириллица, первая буква должна быть заглавной, без пробелов и без цифр, нет спецсимволов (допустим только дефис)',
-  login:
-    'от 3 до 20 символов, латиница, может содержать цифры, но не состоять из них, без пробелов, без спецсимволов (допустимы дефис и нижнее подчёркивание)',
-  email:
-    'латиница, может включать цифры и спецсимволы вроде дефиса, обязательно должна быть «собака» (@) и точка после неё, но перед точкой обязательно должны быть буквы',
-  password:
-    'от 8 до 40 символов, обязательно хотя бы одна заглавная буква и цифра',
-  phone: 'от 10 до 15 символов, состоит из цифр, может начинаться с плюса',
-}
-
-export const regExpByField = {
-  first_name: /^[A-ZА-ЯЁ]{1}[-A-Za-zА-ЯЁа-яё]{2,}$/,
-  second_name: /^[A-ZА-ЯЁ]{1}[-A-Za-zА-ЯЁа-яё]{2,}$/,
-  login: /^[\w\d_-]{3,20}$/,
-  display_name: /^[\wА-ЯЁа-яё\d_-]{3,20}$/,
-  email: /^[\w_-]+@[\w]+[.]{1}[\w]+$/,
-  password: /^[+]*[\d]{10,15}$/,
-  phone: /^[+]*[\d]{10,15}$/,
-}
 
 export const RegistrationPage: FC = () => {
   const { setAuth } = useAuth()
@@ -57,14 +34,7 @@ export const RegistrationPage: FC = () => {
               required: true,
               message: 'Укажите Имя',
             },
-            () => ({
-              validator(_, value) {
-                if (!value || regExpByField.first_name.test(value)) {
-                  return Promise.resolve();
-                }
-                return Promise.reject(new Error('Ошибка валидации поля Имя'));
-              },
-            }),
+            validate(regExpByField.first_name, 'Ошибка валидации поля Имя'),
           ]}
           tooltip={fieldTooltip.first_name}>
           <Input name="first_name" />
@@ -77,14 +47,10 @@ export const RegistrationPage: FC = () => {
               required: true,
               message: 'Укажите Фамилию',
             },
-            () => ({
-              validator(_, value) {
-                if (!value || regExpByField.second_name.test(value)) {
-                  return Promise.resolve();
-                }
-                return Promise.reject(new Error('Ошибка валидации поля Фамилия'));
-              },
-            }),
+            validate(
+              regExpByField.second_name,
+              'Ошибка валидации поля Фамилия'
+            ),
           ]}
           tooltip={fieldTooltip.second_name}>
           <Input name="second_name" />
@@ -97,14 +63,7 @@ export const RegistrationPage: FC = () => {
               required: true,
               message: 'Укажите Логин',
             },
-            () => ({
-              validator(_, value) {
-                if (!value || regExpByField.login.test(value)) {
-                  return Promise.resolve();
-                }
-                return Promise.reject(new Error('Ошибка валидации поля Логин'));
-              },
-            }),
+            validate(regExpByField.login, 'Ошибка валидации поля Логин'),
           ]}
           tooltip={fieldTooltip.login}>
           <Input name="login" autoComplete="on" />
@@ -117,14 +76,7 @@ export const RegistrationPage: FC = () => {
               required: true,
               message: 'Ошибка валидации E-mail!',
             },
-            () => ({
-              validator(_, value) {
-                if (!value || regExpByField.email.test(value)) {
-                  return Promise.resolve();
-                }
-                return Promise.reject(new Error('Ошибка валидации поля Почта'));
-              },
-            }),
+            validate(regExpByField.email, 'Ошибка валидации поля Почта'),
           ]}
           tooltip={fieldTooltip.email}>
           <Input name="email" autoComplete="on" />
@@ -137,14 +89,7 @@ export const RegistrationPage: FC = () => {
               required: true,
               message: 'Укажите Телефон',
             },
-            () => ({
-              validator(_, value) {
-                if (!value || regExpByField.phone.test(value)) {
-                  return Promise.resolve();
-                }
-                return Promise.reject(new Error('Ошибка валидации поля Телефон'));
-              },
-            }),
+            validate(regExpByField.phone, 'Ошибка валидации поля Телефон'),
           ]}
           tooltip={fieldTooltip.phone}>
           <Input name="phone" autoComplete="on" />
@@ -157,14 +102,7 @@ export const RegistrationPage: FC = () => {
               required: true,
               message: 'Укажите Пароль',
             },
-            () => ({
-              validator(_, value) {
-                if (!value || regExpByField.password.test(value)) {
-                  return Promise.resolve();
-                }
-                return Promise.reject(new Error('Ошибка валидации поля Пароль'));
-              },
-            }),
+            validate(regExpByField.password, 'Ошибка валидации поля Пароль'),
           ]}
           tooltip={fieldTooltip.password}>
           <Input.Password name="password" />
@@ -180,9 +118,9 @@ export const RegistrationPage: FC = () => {
             ({ getFieldValue }) => ({
               validator(_, value) {
                 if (!value || getFieldValue('password') === value) {
-                  return Promise.resolve();
+                  return Promise.resolve()
                 }
-                return Promise.reject(new Error('Пароли должны совпадать'));
+                return Promise.reject(new Error('Пароли должны совпадать'))
               },
             }),
           ]}
