+++ packages/client/src/shared/lib/Validation/index.ts
@@ -1,0 +1,34 @@
+export const fieldTooltip = {
+  first_name:
+    'латиница или кириллица, первая буква должна быть заглавной, без пробелов и без цифр, нет спецсимволов (допустим только дефис)',
+  second_name:
+    'латиница или кириллица, первая буква должна быть заглавной, без пробелов и без цифр, нет спецсимволов (допустим только дефис)',
+  login:
+    'от 3 до 20 символов, латиница, может содержать цифры, но не состоять из них, без пробелов, без спецсимволов (допустимы дефис и нижнее подчёркивание)',
+  email:
+    'латиница, может включать цифры и спецсимволы вроде дефиса, обязательно должна быть «собака» (@) и точка после неё, но перед точкой обязательно должны быть буквы',
+  password:
+    'от 8 до 40 символов, обязательно хотя бы одна заглавная буква и цифра',
+  phone: 'от 10 до 15 символов, состоит из цифр, может начинаться с плюса',
+}
+
+export const regExpByField = {
+  first_name: /^[A-ZА-ЯЁ]{1}[-A-Za-zА-ЯЁа-яё]{2,}$/,
+  second_name: /^[A-ZА-ЯЁ]{1}[-A-Za-zА-ЯЁа-яё]{2,}$/,
+  login: /^[\w\d_-]{3,20}$/,
+  display_name: /^[\wА-ЯЁа-яё\d_-]{3,20}$/,
+  email: /^[\w_-]+@[\w]+[.]{1}[\w]+$/,
+  password: /^[+]*[\d]{10,15}$/,
+  phone: /^[+]*[\d]{10,15}$/,
+}
+
+export const validate = (regExp: RegExp, errorMessage: string) => {
+  return () => ({
+    validator(_: unknown, value: string) {
+      if (!value || regExp.test(value)) {
+        return Promise.resolve()
+      }
+      return Promise.reject(new Error(errorMessage))
+    },
+  })
+}
